requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
APScheduler==3.11.2
attrs==26.1.0
bitarray==3.8.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
ckzg==2.1.7
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
cytoolz==1.1.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
fastapi==0.121.2
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.50.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.0.5
langchain-core==1.0.4
langchain-google-genai==3.0.3
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
lxml==5.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
mmh3==5.2.1
MouseInfo==0.1.3
multidict==6.7.1
neat-python==0.92
numpy==2.2.3
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.2
parsimonious==0.10.0
pillow==11.3.0
postgrest==2.28.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
py-solc-x==2.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pymongo==4.15.4
PyMsgBox==2.0.1
pyparsing==3.2.3
pyperclip==1.11.0
PyRect==0.2.0
pyroaring==1.0.4
PyScreeze==1.0.1
pyshark==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
realtime==2.28.3
regex==2026.2.28
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.3.3
rlp==4.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.2
termcolor==2.5.0
toolz==1.1.0
tqdm==4.67.1
types-requests==2.32.4.20260324
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
web3==7.14.1
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
